final
final commit(maybe?)

+++ src/main/java/model/entity/CreditCard.java
@@ -85,7 +85,7 @@ public class CreditCard extends Model {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(long balance) {
         this.balance = balance;
     }
